+++ app/src/main/java/com/demo/weather/weather/data/stations/Data.kt
@@ -5,5 +5,5 @@ import com.google.gson.annotations.SerializedName
 data class Data (
     @SerializedName("id"       ) var id       : String? = null,
     @SerializedName("name"     ) var name     : Name?   = Name(),
-    @SerializedName("distance" ) var distance : Int?    = null
+    @SerializedName("distance" ) var distance : Float?    = null
 )
